requirements.txt
fastapi==0.112.0
uvicorn==0.30.1
asyncpg==0.29.0
openai==1.40.0
httpx==0.27.2
python-dotenv==1.0.1
